backend/foodgram/posts/views.py
from rest_framework import viewsets 
from .models import Recipe, Ingredient
from .serializer import RecipeSerializer, IngredientSerializer


class RecipesViewSet(viewsets.ModelViewSet):
    queryset = Recipe.objects.all()
    serializer_class = RecipeSerializer

class IngredientsViewSet(viewsets.ModelViewSet):
    queryset = Ingredient.objects.all()
    serializer_class = IngredientSerializer